% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_count_levels_removal.R
\name{low_count_levels_removal}
\alias{low_count_levels_removal}
\title{Low Count Factor Levels Removal}
\usage{
low_count_levels_removal(data)
}
\arguments{
\item{data}{Name of data.frame or data.table}
}
\description{
Removes levels of a factor variable with low counts
}
\examples{
low_count_levels_removal()
}
\keyword{factor}
\keyword{levels}

