% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_deciles.R
\name{pareto_deciles}
\alias{pareto_deciles}
\title{Pareto Deciles}
\usage{
pareto_deciles(data, x)
}
\arguments{
\item{data}{Name of data.frame or data.table}

\item{x}{Name of numeric vector to create Pareto deciles for}

\item{id}{Unique ID of data.frame or data.table to group by}
}
\description{
Creates Pareto Deciles of a given metric for each unique ID in your dataset.
}
\examples{
pareto_deciles()
}
\keyword{Pareto}
\keyword{deciles}

