% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified.R
\name{stratified}
\alias{stratified}
\title{Stratified Sampling}
\usage{
stratified(data, group, size, select = NULL, replace = FALSE,
  bothSets = FALSE)
}
\arguments{
\item{data}{Name of data.frame or data.table}

\item{group}{Name of columns to create interaction variables; variable name must be in quotations "", combine multiple columns with c}

\item{size}{Number for the number of rows to keep or a decimal for the percentage of rows to keep}

\item{select}{A named list}

\item{replace}{Should sampling be with replacement?}

\item{bothSets}{Should both sets be used?}
}
\description{
Creates Pareto Deciles of a given metric for each unique ID in your dataset.
}
\examples{
stratified()
}
\keyword{sampling}
\keyword{stratified}

